Fixed a few warnings.
Rev: src/modules/HTTPLoop/accept_and_parse.c:1.19
Rev: src/modules/Image/colortable.c:1.97

+++ src/modules/HTTPLoop/accept_and_parse.c
@@ -335,7 +335,7 @@ void aap_handle_connection(struct args *arg)
   while(1)
 #endif /* HAVE_TIMEOUTS */
   {
-    int data_read = fd_read(arg->fd, p, buffer_len-pos);
+    ptrdiff_t data_read = fd_read(arg->fd, p, buffer_len-pos);
     if(data_read <= 0)
     {
 #ifdef AAP_DEBUG
